+++ amd/periodicset.py
@@ -89,7 +89,7 @@ class PeriodicSet:
 
         cellpar_str = None
         if self.ndim == 1:
-            cellpar_str = f'{self.cell[0][0]}'
+            cellpar_str = f', cell={self.cell[0][0]}'
         if self.ndim == 2:
             cellpar = np.round(cell_to_cellpar_2D(self.cell), 2)
             cellpar_str = ','.join(str(round(p, 2)) for p in cellpar)
@@ -101,36 +101,47 @@ class PeriodicSet:
         else:
             cellpar_str = f''
 
-        s = f'PeriodicSet({self.name}, ' \
+        s = f'PeriodicSet(name={self.name}, ' \
             f'motif {self.motif.shape} ({n_asym_sites} asym sites)' \
             f'{cellpar_str})'
 
         return s
 
-    # def __repr__(self):
-    #     pass
+    def __repr__(self):
 
-    # used for debugging, checks if the motif/cell agree point for point.
+        if self.asymmetric_unit is None:
+            n_asym_sites = len(self.motif)
+        else:
+            n_asym_sites = len(self.asymmetric_unit)
+
+        s = f'PeriodicSet(name={self.name}, ' \
+            f'motif={self.motif} ({n_asym_sites} asym sites), ' \
+            f'cell={self.cell})'
+
+        return s
+
+    
     def __eq__(self, other):
+        """Used for debugging/tests. Returns True if:
+        1. The unit cells are identical (element for element within tol)
+        2. The motifs are the same shape, and when transformed into
+        fractional coordinates every point in one motif has an identical
+        (within tol) point somewhere in the other, accounting for pbc.
+        """
 
-        if self.cell.shape != other.cell.shape or self.motif.shape != other.motif.shape:
+        if self.cell.shape != other.cell.shape or \
+           self.motif.shape != other.motif.shape or \
+           not np.allclose(self.cell, other.cell):
             return False
 
-        if not np.allclose(self.cell, other.cell):
+        fm1 = np.mod(self.motif @ np.linalg.inv(self.cell), 1)
+        fm2 = np.mod(other.motif @ np.linalg.inv(other.cell), 1)
+        d1 = np.abs(fm2[:, None] - fm1)
+        d2 = np.abs(d1 - 1)
+        diffs = np.amax(np.minimum(d1, d2), axis=-1)
+
+        if not np.all((np.amin(diffs, axis=0) <= 1e-8) | (np.amin(diffs, axis=-1) <= 1e-8)):
             return False
-
-        # doesn't even check if motifs are alike because pbcs may make them look different
-
-        # m1 = np.mod(self.motif @ np.linalg.inv(self.cell), 1)
-        # m2 = np.mod(other.motif @ np.linalg.inv(other.cell), 1)
-
-        # diffs = np.amax(np.abs(m2[:, None] - m1), axis=-1)
-        # if not np.all((np.amin(diffs, axis=0) <= 1e-8) | (np.amin(diffs, axis=-1) <= 1e-8)):
-        #     return False
-
-        # diffs = np.amax(np.abs(other.motif[:, None] - self.motif), axis=-1)
-        # if not np.all((np.amin(diffs, axis=0) <= 1e-6) | (np.amin(diffs, axis=-1) <= 1e-6)):
-        #     return False
 
         return True
 
@@ -163,12 +174,21 @@ class PeriodicSet:
         return PeriodicSet(np.zeros((1, dims)), cell)
 
     @staticmethod
-    def _random(n_points, length_bounds=(1, 2), angle_bounds=(60, 120), dims=3):
+    def _random(
+        n_points,
+        length_bounds=(1, 2),
+        angle_bounds=(60, 120),
+        dims=3
+    ):
         """Dimensions 2 and 3 only. Return a :class:`PeriodicSet` with a
         chosen number of randomly placed points, in random cell with
         edges between length_bounds and angles between angle_bounds.
         """
 
-        cell = random_cell(length_bounds=length_bounds, angle_bounds=angle_bounds, dims=dims)
+        cell = random_cell(
+            length_bounds=length_bounds, 
+            angle_bounds=angle_bounds, 
+            dims=dims
+        )
         frac_motif = np.random.uniform(size=(n_points, dims))
         return PeriodicSet(frac_motif @ cell, cell)